Documentation, bugfixing, again and again.

+++ src/main/java/it/polimi/ingsw/lim/controller/rounds/RoundTimer.java
@@ -16,7 +16,7 @@ public class RoundTimer {
      */
     public RoundTimer(int seconds, Round roundCallback){
         timer = new Timer();
-        timer.schedule(new EndTimer(roundCallback), seconds * 1000 /*by default ms (1s = 1000ms)*/);
+        timer.schedule(new EndTimer(roundCallback), (long)(seconds * 1000) /*by default ms (1s = 1000ms)*/);
     }
     private class EndTimer extends TimerTask {
         private Round roundCallback;
